+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- "Simple" API was renamed to SHV and thus `SimpleClient` is now `SHVClient`
+
 ### Fixed
 - Close of TTY client causing `TypeError` to be raised in receive task